edit bank function created

+++ budget/views.py
@@ -5,6 +5,7 @@ from .models import *
 from datetime import datetime
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
+from django.db import transaction
 
 @api_view(['GET'])
 def home(request):
@@ -16,7 +17,7 @@ def create_debt(request):
     if request.method == 'POST':
         data = request.data
 
-        required_fields = ['bank', 'value', 'maturity', 'month' ,'id_responsible']
+        required_fields = ['id_bank', 'value', 'maturity', 'month' ,'id_responsible']
         for field in required_fields:
             if field not in data:
                 return Response({f'{field} é obrigatório'}, status=status.HTTP_400_BAD_REQUEST)
@@ -32,9 +33,13 @@ def create_debt(request):
             id_responsible = data['id_responsible']
             responsible = Responsible.objects.get(id=id_responsible)
 
+            # Certifica-se de que o Responsible com o ID fornecido exist
+            id_bank = data['id_bank']
+            bank= Bank.objects.get(id=id_bank)
+
             # Cria a instância de Debts associada ao Responsible
             debt = Debts.objects.create(
-                bank=data['bank'],
+                id_bank=bank,
                 value=value,
                 maturity=maturity,
                 month=data['month'],
@@ -62,12 +67,19 @@ def get_debts(request):
         'debts': [
             {
                 'id': debt.id,
-                'bank': debt.bank,
                 'value': debt.value,
                 'maturity': debt.maturity.strftime('%Y-%m-%d'),
                 'month':debt.month,
-                'id_responsible': debt.id_responsible.id,
-
+                'bank': {
+                    'id': debt.id_bank.id,
+                    'nome': debt.id_bank.name,
+                    'cnpj': debt.id_bank.cnpj,
+                    'digital_bank': debt.id_bank.digital_bank,
+                },
+                'responsible' : {
+                    'id': debt.id_responsible.id,
+                    'name': debt.id_responsible.name
+                },
             }
             for debt in debts
         ],
@@ -102,13 +114,15 @@ def create_responsibles(request):
 
 @api_view(['DELETE'])
 def delete_debt(request, id):
-    # Obtém a instância da dívida com base no ID fornecido
-    debt = get_object_or_404(Debts, id=id)
-
-    # Deleta a instância da dívida
-    debt.delete()
-
-    return Response({'message': 'Dívida deletada com sucesso'}, status=status.HTTP_204_NO_CONTENT)
+    if request.method == 'DELETE':
+        try: 
+            debt = Debts.objects.get(id=id)
+            # Deleta a instância da dívida
+            debt.delete()
+            return Response({'message': 'Dívida deletada com sucesso'}, status=status.HTTP_204_NO_CONTENT)
+        except Debts.DoesNotExist:
+            return Response({'message': 'Essa dívida não existe'}, status=status.HTTP_400_BAD_REQUEST)
+    return Response({'message': 'Método não permitido'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
 @api_view(['POST'])
 def create_bank(request):
@@ -152,3 +166,25 @@ def get_banks(request):
 
         return JsonResponse(jsonBanks)
     return Response({'error': f'Método {request.method} não permitido'}, status=status.HTTP_400_BAD_REQUEST)
+
+@api_view(['PUT'])
+def edit_bank(request):
+    bank_id = request.GET.get('id')
+    try:
+        bank = Bank.objects.get(id=bank_id)
+
+        data = request.data
+
+        bank.name = data['name']
+        bank.cnpj = data['cnpj']
+        bank.digital_bank = data['digital_bank']
+
+        bank.save()
+
+        result = {'result': f'Banco {bank.name} editado com sucesso!'}
+
+        return Response(result)
+    except Bank.DoesNotExist:
+        return Response({'result': f'Banco com id {bank_id} não existe!'}, status=status.HTTP_404_NOT_FOUND)
+    except:
+        return Response({'result': 'Um ou mais paramêtros não foram encontrados na requisição!'}, status=status.HTTP_406_NOT_ACCEPTABLE)
